kw-tester: fix convexPolygon arbitrary — fc.float max must be Math.fround
fast-check v4 requires float constraints to be 32-bit floats.
Math.PI * 2 is a 64-bit double; wrapping in Math.fround makes it valid.
Fixes a latent crash when any test imports convexPolygon(n).

+++ tests/arbitraries.ts
@@ -271,7 +271,10 @@ export function convexPolygon(n: number): fc.Arbitrary<Point2D[]> {
   if (n < 3) throw new Error("convexPolygon requires n >= 3");
   return fc
     .array(
-      fc.tuple(fc.float({ min: 0, max: Math.PI * 2, noNaN: true }), fc.integer({ min: 1, max: 5 })),
+      fc.tuple(
+        fc.float({ min: 0, max: Math.fround(Math.PI * 2), noNaN: true }),
+        fc.integer({ min: 1, max: 5 }),
+      ),
       { minLength: n, maxLength: n },
     )
     .map((pairs) => {
